api/requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
anthropic
slowapi
redis
requests
beautifulsoup4